[#108] rename boot => core
