+++ rupring/src/core/banner.rs
